+++ lib/edgedb/protocol.ex
@@ -16,7 +16,7 @@ defmodule EdgeDB.Protocol do
   }
 
   @type message_code() :: pos_integer()
-  @type protocol_version() :: non_neg_integer()
+  @type protocol_version() :: {non_neg_integer(), non_neg_integer()}
 
   @type client_message() ::
           Client.V0.Flush.t()
@@ -54,6 +54,49 @@ defmodule EdgeDB.Protocol do
 
   @message_header_length 5
 
+  @major_version 2
+  @minor_version 0
+
+  @min_major_version 0
+  @min_minor_version 13
+
+  @set 0
+  @output_shape 1
+  @base_scalar 2
+  @scalar 3
+  @tuple 4
+  @named_tuple 5
+  @array 6
+  @enum 7
+  @input_shape 8
+  @range 9
+  @object 10
+  @compound 11
+  @type_annotation_text 127
+  @type_annotation_min 0x80
+  @type_annotatin_max 0xFE
+  @scalar_type_name_annotation 0xFF
+
+  @spec major_version() :: unquote(@major_version)
+  def major_version do
+    @major_version
+  end
+
+  @spec minor_version() :: unquote(@minor_version)
+  def minor_version do
+    @minor_version
+  end
+
+  @spec min_major_version() :: unquote(@min_major_version)
+  def min_major_version do
+    @min_major_version
+  end
+
+  @spec min_minor_version() :: unquote(@min_minor_version)
+  def min_minor_version do
+    @min_minor_version
+  end
+
   @spec message_header_length() :: unquote(@message_header_length)
   def message_header_length do
     @message_header_length
@@ -90,16 +133,16 @@ defmodule EdgeDB.Protocol do
     decode_message(type, message_payload, protocol_version)
   end
 
-  @spec parse_type_description(bitstring(), CodecStorage.t()) :: Codec.id()
-  def parse_type_description(type_description, codec_storage) do
-    do_type_description_parsing(type_description, codec_storage, %{})
+  @spec parse_type_description(bitstring(), CodecStorage.t(), protocol_version()) :: Codec.id()
+  def parse_type_description(type_description, codec_storage, protocol) do
+    do_type_description_parsing(type_description, codec_storage, %{}, protocol)
   end
 
-  defp do_message_encoding(%Client.V0.Flush{}, 0) do
+  defp do_message_encoding(%Client.V0.Flush{}, {0, _minor}) do
     {0x48, []}
   end
 
-  defp do_message_encoding(%Client.V0.Prepare{} = message, 0) do
+  defp do_message_encoding(%Client.V0.Prepare{} = message, {0, _minor}) do
     headers = map_into_key_value_list(:v0_prepare, message.headers)
 
     {0x50,
@@ -113,7 +156,7 @@ defmodule EdgeDB.Protocol do
      ]}
   end
 
-  defp do_message_encoding(%Client.V0.DescribeStatement{} = message, 0) do
+  defp do_message_encoding(%Client.V0.DescribeStatement{} = message, {0, _minor}) do
     headers = map_into_key_value_list(:v0_describe_statement, message.headers)
 
     {0x44,
@@ -124,7 +167,7 @@ defmodule EdgeDB.Protocol do
      ]}
   end
 
-  defp do_message_encoding(%Client.V0.Execute{} = message, 0) do
+  defp do_message_encoding(%Client.V0.Execute{} = message, {0, _minor}) do
     headers = map_into_key_value_list(:v0_execute, message.headers)
 
     {0x45,
@@ -135,7 +178,7 @@ defmodule EdgeDB.Protocol do
      ]}
   end
 
-  defp do_message_encoding(%Client.V0.OptimisticExecute{} = message, 0) do
+  defp do_message_encoding(%Client.V0.OptimisticExecute{} = message, {0, _minor}) do
     headers = map_into_key_value_list(:v0_optimisitc_execute, message.headers)
 
     {0x4F,
@@ -151,7 +194,7 @@ defmodule EdgeDB.Protocol do
      ]}
   end
 
-  defp do_message_encoding(%Client.V0.ExecuteScript{} = message, 0) do
+  defp do_message_encoding(%Client.V0.ExecuteScript{} = message, {0, _minor}) do
     headers = map_into_key_value_list(:v0_execute_script, message.headers)
 
     {0x51,
@@ -181,7 +224,7 @@ defmodule EdgeDB.Protocol do
   end
 
   # EdgeDB 1.0 doesn't define any extensions, so it's safe to encode them as empty list
-  defp do_message_encoding(%Client.ClientHandshake{} = message, 0) do
+  defp do_message_encoding(%Client.ClientHandshake{} = message, {0, _minor}) do
     {0x56,
      [
        <<message.major_ver::uint16(), message.minor_ver::uint16()>>,
@@ -246,7 +289,7 @@ defmodule EdgeDB.Protocol do
     {0x58, []}
   end
 
-  defp do_message_decoding(0x31, <<data::binary>>, 0) do
+  defp do_message_decoding(0x31, <<data::binary>>, {0, _minor}) do
     {headers,
      <<
        cardinality::uint8(),
@@ -291,7 +334,7 @@ defmodule EdgeDB.Protocol do
     %Server.AuthenticationSASLFinal{auth_status: 0xB, sasl_data: sasl_data}
   end
 
-  defp do_message_decoding(0x43, <<data::binary>>, 0) do
+  defp do_message_decoding(0x43, <<data::binary>>, {0, _minor}) do
     {headers, <<status_size::uint32(), status::binary(status_size)>>} =
       decode_key_value_list(data)
 
@@ -322,7 +365,7 @@ defmodule EdgeDB.Protocol do
     }
   end
 
-  defp do_message_decoding(0x54, <<data::binary>>, 0) do
+  defp do_message_decoding(0x54, <<data::binary>>, {0, _minor}) do
     {headers,
      <<
        result_cardinality::uint8(),
@@ -397,7 +440,7 @@ defmodule EdgeDB.Protocol do
     }
   end
 
-  defp do_message_decoding(0x4C, <<severity::uint8(), code::uint32(), rest::binary>>, 0) do
+  defp do_message_decoding(0x4C, <<severity::uint8(), code::uint32(), rest::binary>>, {0, _minor}) do
     <<text_size::uint32(), text::binary(text_size), rest::binary>> = rest
     {annotations, <<>>} = decode_key_value_list(rest)
 
@@ -434,7 +477,7 @@ defmodule EdgeDB.Protocol do
     %Server.ParameterStatus{name: name, value: decode_parameter_status_value(name, value)}
   end
 
-  defp do_message_decoding(0x5A, <<data::binary>>, 0) do
+  defp do_message_decoding(0x5A, <<data::binary>>, {0, _minor}) do
     {headers, <<transaction_state::uint8()>>} = decode_key_value_list(data)
 
     %Server.ReadyForCommand{
@@ -453,7 +496,11 @@ defmodule EdgeDB.Protocol do
   end
 
   # EdgeDB 1.0 doesn't define any extensions, so it's safe to decode them as empty list
-  defp do_message_decoding(0x76, <<major_ver::uint16(), minor_ver::uint16(), 0::uint16()>>, 0) do
+  defp do_message_decoding(
+         0x76,
+         <<major_ver::uint16(), minor_ver::uint16(), 0::uint16()>>,
+         {0, _minor}
+       ) do
     %Server.ServerHandshake{
       major_ver: major_ver,
       minor_ver: minor_ver,
@@ -480,169 +527,262 @@ defmodule EdgeDB.Protocol do
   end
 
   defp do_type_description_parsing(
-         <<type::uint8(), id::uuid(), data::binary>>,
+         <<
+           _desc_size::uint32(),
+           @type_annotation_text::uint8(),
+           _descriptor::uint16(),
+           key_size::uint32(),
+           _key::binary(key_size),
+           value_size::uint32(),
+           _value::binary(value_size),
+           rest::binary
+         >>,
          codec_storage,
-         codecs
-       ) do
-    rest =
-      case CodecStorage.get(codec_storage, id) do
-        nil ->
-          {codec, rest} = do_codec_parsing(type, data, id, codecs, true)
-          CodecStorage.add(codec_storage, id, codec)
-          rest
-
-        _codec ->
-          {_codec, rest} = do_codec_parsing(type, data, id, codecs, false)
-          rest
-      end
-
-    do_type_description_parsing(rest, codec_storage, Map.put(codecs, map_size(codecs), id))
+         codecs,
+         {major, _minor} = protocol
+       )
+       when major >= 2 do
+    do_type_description_parsing(rest, codec_storage, codecs, protocol)
   end
 
-  defp do_type_description_parsing(<<>>, _codec_storage, codecs) do
+  defp do_type_description_parsing(
+         <<
+           type::uint8(),
+           _id::uuid(),
+           annotation_size::uint32(),
+           _annotation::binary(annotation_size),
+           rest::binary
+         >>,
+         codec_storage,
+         codecs,
+         {major, _minor} = protocol
+       )
+       when type in @type_annotation_min..@type_annotatin_max and major <= 1 do
+    do_type_description_parsing(rest, codec_storage, codecs, protocol)
+  end
+
+  defp do_type_description_parsing(
+         <<
+           @scalar_type_name_annotation::uint8(),
+           id::uuid(),
+           type_name_size::uint32(),
+           type_name::binary(type_name_size),
+           rest::binary
+         >>,
+         codec_storage,
+         codecs,
+         {major, _minor} = protocol
+       )
+       when major <= 1 do
+    with %codec_name{} = codec when codec_name in [Codecs.Scalar, Codecs.Enum] <-
+           CodecStorage.get(codec_storage, id) do
+      codec = %{codec | name: type_name}
+      CodecStorage.add(codec_storage, id, codec)
+    end
+
+    do_type_description_parsing(rest, codec_storage, codecs, protocol)
+  end
+
+  defp do_type_description_parsing(
+         <<desc_size::uint32(), descriptor::binary(desc_size), rest::binary>>,
+         codec_storage,
+         codecs,
+         {major, _minor} = protocol
+       )
+       when major >= 2 do
+    <<type::uint8(), id::uuid(), data::binary>> = descriptor
+
+    with nil <- CodecStorage.get(codec_storage, id) do
+      {codec, <<>>} = do_codec_parsing(type, data, id, codecs, protocol)
+      CodecStorage.add(codec_storage, id, codec)
+    end
+
+    do_type_description_parsing(
+      rest,
+      codec_storage,
+      Map.put(codecs, map_size(codecs), id),
+      protocol
+    )
+  end
+
+  defp do_type_description_parsing(
+         <<type::uint8(), id::uuid(), data::binary>>,
+         codec_storage,
+         codecs,
+         protocol
+       ) do
+    {codec, rest} = do_codec_parsing(type, data, id, codecs, protocol)
+
+    if is_nil(CodecStorage.get(codec_storage, id)) do
+      CodecStorage.add(codec_storage, id, codec)
+    end
+
+    do_type_description_parsing(
+      rest,
+      codec_storage,
+      Map.put(codecs, map_size(codecs), id),
+      protocol
+    )
+  end
+
+  defp do_type_description_parsing(<<>>, _codec_storage, codecs, _protocol) do
     codecs[map_size(codecs) - 1]
   end
 
-  defp do_codec_parsing(0, <<type_pos::uint16(), rest::binary>>, id, codecs, create?) do
-    codec =
-      if create? do
-        Codecs.Set.new(id, codecs[type_pos])
-      else
-        nil
-      end
-
-    {codec, rest}
+  defp do_codec_parsing(@set, <<type_pos::uint16(), rest::binary>>, id, codecs, _protocol) do
+    {Codecs.Set.new(id, codecs[type_pos]), rest}
   end
 
-  defp do_codec_parsing(1, <<data::binary>>, id, codecs, create?) do
-    {shape_elements, rest} = decode_shape_element_list(data)
+  defp do_codec_parsing(
+         @output_shape,
+         <<_ephemeral::bool(), _type::uint16(), data::binary>>,
+         id,
+         codecs,
+         {major, _minor} = protocol
+       )
+       when major >= 2 do
+    {shape_elements, rest} = decode_output_shape_element_list(data, protocol)
     codecs = Enum.map(shape_elements, &codecs[&1.type_pos])
 
-    codec =
-      if create? do
-        Codecs.Object.new(id, shape_elements, codecs, false)
-      else
-        nil
-      end
-
-    {codec, rest}
+    {Codecs.Object.new(id, nil, shape_elements, codecs, false), rest}
   end
 
-  defp do_codec_parsing(2, <<rest::binary>>, _id, _codecs, false) do
+  defp do_codec_parsing(@output_shape, <<data::binary>>, id, codecs, protocol) do
+    {shape_elements, rest} = decode_output_shape_element_list(data, protocol)
+    codecs = Enum.map(shape_elements, &codecs[&1.type_pos])
+
+    {Codecs.Object.new(id, nil, shape_elements, codecs, false), rest}
+  end
+
+  defp do_codec_parsing(@base_scalar, <<rest::binary>>, _id, _codecs, _protocol) do
     {nil, rest}
   end
 
-  defp do_codec_parsing(3, <<base_type_pos::uint16(), rest::binary>>, id, codecs, create?) do
-    codec =
-      if create? do
-        Codecs.Scalar.new(id, codecs[base_type_pos])
-      else
-        nil
-      end
+  defp do_codec_parsing(@scalar, <<data::binary>>, id, codecs, {major, _minor})
+       when major >= 2 do
+    <<name_size::uint32(), name::binary(name_size), rest::binary>> = data
+    <<_schema_defined::bool(), rest::binary>> = rest
+    {ancestors, rest} = decode_uint16_list(rest)
 
-    {codec, rest}
+    base_type_idx = List.last(ancestors)
+
+    {Codecs.Scalar.new(id, name, codecs[base_type_idx]), rest}
   end
 
-  defp do_codec_parsing(4, <<data::binary>>, id, codecs, create?) do
+  defp do_codec_parsing(@scalar, <<base_type_pos::uint16(), rest::binary>>, id, codecs, _protocol) do
+    {Codecs.Scalar.new(id, nil, codecs[base_type_pos]), rest}
+  end
+
+  defp do_codec_parsing(@tuple, <<data::binary>>, id, codecs, {major, _minor}) when major >= 2 do
+    <<name_size::uint32(), name::binary(name_size), rest::binary>> = data
+    <<_schema_defined::bool(), rest::binary>> = rest
+    {_ancestors, rest} = decode_uint16_list(rest)
+
+    {element_types, rest} = decode_uint16_list(rest)
+    codecs = Enum.map(element_types, &codecs[&1])
+
+    {Codecs.Tuple.new(id, name, codecs), rest}
+  end
+
+  defp do_codec_parsing(@tuple, <<data::binary>>, id, codecs, _protocol) do
     {element_types, rest} = decode_uint16_list(data)
     codecs = Enum.map(element_types, &codecs[&1])
 
-    codec =
-      if create? do
-        Codecs.Tuple.new(id, codecs)
-      else
-        nil
-      end
-
-    {codec, rest}
+    {Codecs.Tuple.new(id, nil, codecs), rest}
   end
 
-  defp do_codec_parsing(5, <<data::binary>>, id, codecs, create?) do
+  defp do_codec_parsing(@named_tuple, <<data::binary>>, id, codecs, {major, _minor})
+       when major >= 2 do
+    <<name_size::uint32(), name::binary(name_size), rest::binary>> = data
+    <<_schema_defined::bool(), rest::binary>> = rest
+    {_ancestors, rest} = decode_uint16_list(rest)
+
+    {elements, rest} = decode_tuple_element_list(rest)
+    codecs = Enum.map(elements, &codecs[&1.type_pos])
+
+    {Codecs.NamedTuple.new(id, name, elements, codecs), rest}
+  end
+
+  defp do_codec_parsing(@named_tuple, <<data::binary>>, id, codecs, _protocol) do
     {elements, rest} = decode_tuple_element_list(data)
     codecs = Enum.map(elements, &codecs[&1.type_pos])
 
-    codec =
-      if create? do
-        Codecs.NamedTuple.new(id, elements, codecs)
-      else
-        nil
-      end
-
-    {codec, rest}
+    {Codecs.NamedTuple.new(id, nil, elements, codecs), rest}
   end
 
-  defp do_codec_parsing(6, <<type_pos::uint16(), rest::binary>>, id, codecs, create?) do
+  defp do_codec_parsing(@array, <<data::binary>>, id, codecs, {major, _minor})
+       when major >= 2 do
+    <<name_size::uint32(), name::binary(name_size), rest::binary>> = data
+    <<_schema_defined::bool(), rest::binary>> = rest
+    {_ancestors, rest} = decode_uint16_list(rest)
+
+    <<type::uint16(), rest::binary>> = rest
     {dimensions, rest} = decode_int32_list(rest)
 
-    codec =
-      if create? do
-        Codecs.Array.new(id, codecs[type_pos], dimensions)
-      else
-        nil
-      end
-
-    {codec, rest}
+    {Codecs.Array.new(id, name, codecs[type], dimensions), rest}
   end
 
-  defp do_codec_parsing(7, <<data::binary>>, id, _codecs, create?) do
+  defp do_codec_parsing(@array, <<type_pos::uint16(), rest::binary>>, id, codecs, _protocol) do
+    {dimensions, rest} = decode_int32_list(rest)
+
+    {Codecs.Array.new(id, nil, codecs[type_pos], dimensions), rest}
+  end
+
+  defp do_codec_parsing(@enum, <<data::binary>>, id, _codecs, {major, _minor}) when major >= 2 do
+    <<name_size::uint32(), name::binary(name_size), rest::binary>> = data
+    <<_schema_defined::bool(), rest::binary>> = rest
+    {_ancestors, rest} = decode_uint16_list(rest)
+
+    {members, rest} = decode_string_list(rest)
+
+    {Codecs.Enum.new(id, name, members), rest}
+  end
+
+  defp do_codec_parsing(@enum, <<data::binary>>, id, _codecs, _protocol) do
     {members, rest} = decode_string_list(data)
 
-    codec =
-      if create? do
-        Codecs.Enum.new(id, members)
-      else
-        nil
-      end
-
-    {codec, rest}
+    {Codecs.Enum.new(id, nil, members), rest}
   end
 
-  defp do_codec_parsing(8, <<data::binary>>, id, codecs, create?) do
-    {shape_elements, rest} = decode_shape_element_list(data)
+  defp do_codec_parsing(@input_shape, <<data::binary>>, id, codecs, _protocol) do
+    {shape_elements, rest} = decode_input_shape_element(data)
     codecs = Enum.map(shape_elements, &codecs[&1.type_pos])
 
-    codec =
-      if create? do
-        Codecs.Object.new(id, shape_elements, codecs, true)
-      else
-        nil
-      end
-
-    {codec, rest}
+    {Codecs.Object.new(id, nil, shape_elements, codecs, true), rest}
   end
 
-  defp do_codec_parsing(9, <<type_pos::uint16(), rest::binary>>, id, codecs, create?) do
+  defp do_codec_parsing(@range, <<data::binary>>, id, codecs, {major, _minor})
+       when major >= 2 do
+    <<name_size::uint32(), name::binary(name_size), rest::binary>> = data
+    <<_schema_defined::bool(), rest::binary>> = rest
+    {_ancestors, rest} = decode_uint16_list(rest)
+
+    <<type::uint16(), rest::binary>> = rest
+    sub_codec = codecs[type]
+
+    {Codecs.Range.new(id, name, sub_codec), rest}
+  end
+
+  defp do_codec_parsing(@range, <<type_pos::uint16(), rest::binary>>, id, codecs, _protocol) do
     sub_codec = codecs[type_pos]
 
-    codec =
-      if create? do
-        Codecs.Range.new(id, sub_codec)
-      else
-        nil
-      end
-
-    {codec, rest}
+    {Codecs.Range.new(id, nil, sub_codec), rest}
   end
 
-  defp do_codec_parsing(
-         0xFF,
-         <<type_name_size::uint32(), _type_name::binary(type_name_size), rest::binary>>,
-         _id,
-         _codecs,
-         false
-       ) do
-    {nil, rest}
+  defp do_codec_parsing(@object, <<data::binary>>, _id, _codecs, _protocol) do
+    <<name_size::uint32(), _name::binary(name_size), rest::binary>> = data
+    <<_schema_defined::bool(), rest::binary>> = rest
+
+    {Codecs.Null.new(), rest}
   end
 
-  defp do_codec_parsing(
-         type,
-         <<annotation_size::uint32(), _annotation::binary(annotation_size), rest::binary>>,
-         _id,
-         _codecs,
-         false
-       )
-       when type in 0x80..0xFE do
-    {nil, rest}
+  defp do_codec_parsing(@compound, <<data::binary>>, _id, _codecs, _protocol) do
+    <<name_size::uint32(), _name::binary(name_size), rest::binary>> = data
+    <<_schema_defined::bool(), op::uint8(), rest::binary>> = rest
+    _op = decode_enum(:type_operation, op)
+    {_components, rest} = decode_uint16_list(rest)
+
+    {Codecs.Null.new(), rest}
   end
 
   defp decode_parameter_status_value("system_config", data) do
@@ -808,15 +948,43 @@ defmodule EdgeDB.Protocol do
     decode_string_list(rest, count - 1, [string | acc])
   end
 
-  defp decode_shape_element_list(<<element_count::uint16(), data::binary>>) do
-    decode_shape_element_list(data, element_count, [])
+  defp decode_output_shape_element_list(<<element_count::uint16(), data::binary>>, protocol) do
+    decode_output_shape_element_list(data, element_count, [], protocol)
   end
 
-  defp decode_shape_element_list(<<data::binary>>, 0, acc) do
+  defp decode_output_shape_element_list(<<data::binary>>, 0, acc, _protocol) do
     {Enum.reverse(acc), data}
   end
 
-  defp decode_shape_element_list(<<data::binary>>, count, acc) do
+  defp decode_output_shape_element_list(<<data::binary>>, count, acc, {major, _minor} = protocol)
+       when major >= 2 do
+    <<
+      flags::uint32(),
+      cardinality::uint8(),
+      name_size::uint32(),
+      name::binary(name_size),
+      type::uint16(),
+      _source_type_pos::uint16(),
+      rest::binary
+    >> = data
+
+    decode_output_shape_element_list(
+      rest,
+      count - 1,
+      [
+        %Types.ShapeElement{
+          flags: flags,
+          cardinality: decode_enum(:cardinality, cardinality),
+          name: name,
+          type_pos: type
+        }
+        | acc
+      ],
+      protocol
+    )
+  end
+
+  defp decode_output_shape_element_list(<<data::binary>>, count, acc, protocol) do
     <<
       flags::uint32(),
       cardinality::uint8(),
@@ -826,7 +994,41 @@ defmodule EdgeDB.Protocol do
       rest::binary
     >> = data
 
-    decode_shape_element_list(rest, count - 1, [
+    decode_output_shape_element_list(
+      rest,
+      count - 1,
+      [
+        %Types.ShapeElement{
+          flags: flags,
+          cardinality: decode_enum(:cardinality, cardinality),
+          name: name,
+          type_pos: type_pos
+        }
+        | acc
+      ],
+      protocol
+    )
+  end
+
+  defp decode_input_shape_element(<<element_count::uint16(), data::binary>>) do
+    decode_input_shape_element(data, element_count, [])
+  end
+
+  defp decode_input_shape_element(<<data::binary>>, 0, acc) do
+    {Enum.reverse(acc), data}
+  end
+
+  defp decode_input_shape_element(<<data::binary>>, count, acc) do
+    <<
+      flags::uint32(),
+      cardinality::uint8(),
+      name_size::uint32(),
+      name::binary(name_size),
+      type_pos::uint16(),
+      rest::binary
+    >> = data
+
+    decode_input_shape_element(rest, count - 1, [
       %Types.ShapeElement{
         flags: flags,
         cardinality: decode_enum(:cardinality, cardinality),
@@ -1112,6 +1314,14 @@ defmodule EdgeDB.Protocol do
 
   defp decode_enum(:transaction_state, 0x45) do
     :in_failed_transaction
+  end
+
+  defp decode_enum(:type_operation, 1) do
+    :union
+  end
+
+  defp decode_enum(:type_operation, 2) do
+    :intersection
   end
 
   defp process_capabilities(capability) when is_atom(capability) do