+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - rendering hints for query errors from EdgeDB.
+- support for EdgeDB binary protocol `2.0`.
 
 ## [0.6.1] - 2023-07-07